requirements.txt
fastapi[all]
pyserial
pydantic
